Model defproperty (#180)
* User-definable properties.

* Simplify parse failure handling in analysis.

* Refactor some of parser to be easier to read.

* Attach parser names in correct positions.

* Allow user-defined properties to take args.

* Remove duplicated `parseBindings`.

* Add a docs for `parseBindings` and `parseDefprops`.

* Tweak `parseToPreProp`.

* Split `Pact.Analyze.Parse` into three modules.

* Make sure property constants can't refer to themselves.

* Add note about inlining.
