+++ src-ghc/Pact/PersistPactDb/Regression.hs
@@ -41,10 +41,10 @@ runRegression p = do
   let ks = KeySet [PublicKey "skdjhfskj"] (Name "predfun" def)
   _writeRow pactdb Write KeySets "ks1" ks v
   assertEquals' "keyset write" (Just ks) $ _readRow pactdb KeySets "ks1" v
-  let mod' = Module "mod1" "mod-admin-keyset" (Meta Nothing M.empty) "code" (H.hash "code") mempty
+  let mod' = Module "mod1" "mod-admin-keyset" (Meta Nothing Nothing) "code" (H.hash "code") mempty
   _writeRow pactdb Write Modules "mod1" mod' v
   assertEquals' "module write" (Just mod') $ _readRow pactdb Modules "mod1" v
-  assertEquals' "hash of commit 3" 5222047446822369474 (hashWithSalt 1 <$> commit v)
+  assertEquals' "hash of commit 3" (-8810669039792530227) (hashWithSalt 1 <$> commit v)
   _t4 <- begin v t3
   tids <- _txids pactdb user1 t1 v
   assertEquals "user txids" [2] tids
