+++ src/Pact/Analyze/Check.hs
@@ -1,8 +1,10 @@
-{-# LANGUAGE LambdaCase          #-}
-{-# LANGUAGE NamedFieldPuns      #-}
-{-# LANGUAGE OverloadedStrings   #-}
-{-# LANGUAGE ScopedTypeVariables #-}
-{-# LANGUAGE TupleSections       #-}
+{-# LANGUAGE LambdaCase            #-}
+{-# LANGUAGE MultiParamTypeClasses #-}
+{-# LANGUAGE NamedFieldPuns        #-}
+{-# LANGUAGE OverloadedStrings     #-}
+{-# LANGUAGE PatternSynonyms       #-}
+{-# LANGUAGE ScopedTypeVariables   #-}
+{-# LANGUAGE TupleSections         #-}
 
 module Pact.Analyze.Check
   ( verifyModule
@@ -10,6 +12,7 @@ module Pact.Analyze.Check
   , describeCheckFailure
   , describeCheckResult
   , describeParseFailure
+  , describeVerificationWarnings
   , showModel
   , CheckFailure(..)
   , CheckFailureNoLoc(..)
@@ -22,17 +25,17 @@ module Pact.Analyze.Check
   ) where
 
 import           Control.Exception         as E
-import           Control.Lens              (ifoldrM, itraversed, ix, traversed,
-                                            view, (<&>), (^.), (^?), (^@..),
-                                            _1, _2)
-import           Control.Monad             (join, void)
-import           Control.Monad.Except      (ExceptT (ExceptT), MonadError,
-                                            catchError, runExceptT, throwError,
-                                            withExcept, withExceptT)
+import           Control.Lens              (at, ifoldrM, itraversed, ix,
+                                            traversed, view, (%~), (&), (<&>),
+                                            (^.), (^?), (^@..), _1, _2, _Just,
+                                            _Left)
+import           Control.Monad             (void)
+import           Control.Monad.Except      (Except, ExceptT (ExceptT),
+                                            MonadError, catchError, runExceptT,
+                                            throwError, withExcept, withExceptT)
 import           Control.Monad.Morph       (generalize, hoist)
 import           Control.Monad.Reader      (runReaderT)
 import           Control.Monad.Trans.Class (MonadTrans (lift))
-import           Data.Either               (lefts, partitionEithers, rights)
 import qualified Data.HashMap.Strict       as HM
 import           Data.Map.Strict           (Map)
 import qualified Data.Map.Strict           as Map
@@ -56,12 +59,12 @@ import           Pact.Types.Lang           (Code (Code), Info (Info), eParsed,
 import           Pact.Types.Runtime        (Exp, ModuleData, ModuleName,
                                             Ref (Ref),
                                             Term (TConst, TDef, TSchema,
-                                            TTable), asString, tShow)
+                                            TTable), asString, tShow, pattern ELitList, pattern EAtom', pattern EList')
 import qualified Pact.Types.Runtime        as Pact
 import           Pact.Types.Typecheck      (AST,
                                             Fun (FDefun, _fArgs, _fBody, _fInfo),
                                             Named, Node, TcId (_tiInfo),
-                                            TopLevel (TopFun, TopTable, _tlInfo),
+                                            TopLevel (TopFun, TopTable),
                                             UserType (_utFields, _utName),
                                             runTC, tcFailures)
 import qualified Pact.Types.Typecheck      as TC
@@ -70,10 +73,19 @@ import           Pact.Analyze.Errors
 import           Pact.Analyze.Eval         hiding (invariants)
 import           Pact.Analyze.Model        (allocArgs, allocModelTags,
                                             saturateModel, showModel)
-import           Pact.Analyze.Parse        (expToCheck, expToInvariant)
+import           Pact.Analyze.Parse        (expToCheck, expToInvariant,
+                                            parseBindings)
 import           Pact.Analyze.Translate
 import           Pact.Analyze.Types
 import           Pact.Analyze.Util
+
+newtype VerificationWarnings = VerificationWarnings [Text]
+  deriving (Eq, Show)
+
+describeVerificationWarnings :: VerificationWarnings -> Text
+describeVerificationWarnings (VerificationWarnings dups) = case dups of
+  [] -> ""
+  _  -> "Warning: duplicated property definitions for " <> T.intercalate ", " dups
 
 data CheckSuccess
   = SatisfiedProperty Model
@@ -115,10 +127,11 @@ type CheckResult = Either CheckFailure CheckSuccess
 data ModuleChecks = ModuleChecks
   { propertyChecks  :: HM.HashMap Text [CheckResult]
   , invariantChecks :: HM.HashMap Text (TableMap [CheckResult])
+  , moduleWarnings  :: VerificationWarnings
   } deriving (Eq, Show)
 
 data VerificationFailure
-  = ModuleParseFailures [ParseFailure]
+  = ModuleParseFailure ParseFailure
   | ModuleCheckFailure CheckFailure
   | TypeTranslationFailure Text (Pact.Type TC.UserType)
   deriving Show
@@ -314,8 +327,8 @@ verifyFunctionProperty funInfo tables pactArgs body (Located propInfo check) =
 moduleTables
   :: HM.HashMap ModuleName ModuleData -- ^ all loaded modules
   -> ModuleData                       -- ^ the module we're verifying
-  -> ExceptT [ParseFailure] IO [Table]
-moduleTables modules (_mod, modRefs) = ExceptT $ do
+  -> ExceptT ParseFailure IO [Table]
+moduleTables modules (_mod, modRefs) = do
   -- All tables defined in this module, and imported by it. We're going to look
   -- through these for their schemas, which we'll look through for invariants.
   let tables = flip mapMaybe (modules ^@.. traversed . _2 . itraversed) $ \case
@@ -327,25 +340,41 @@ moduleTables modules (_mod, modRefs) = ExceptT $ do
         (name, Ref (schema@TSchema {})) -> Just (name, schema)
         _                               -> Nothing
 
-  eitherTables <- for tables $ \(tabName, tab) -> do
-      (TopTable _info _name (Pact.TyUser schema) _meta, _tcState)
-        <- runTC 0 False $ typecheckTopLevel (Ref tab)
+  for tables $ \(tabName, tab) -> do
+    (TopTable _info _name (Pact.TyUser schema) _meta, _tcState)
+      <- lift $ runTC 0 False $ typecheckTopLevel (Ref tab)
 
-      let TC.Schema{_utName,_utFields} = schema
-          schemaName = asString _utName
+    let TC.Schema{_utName,_utFields} = schema
+        schemaName = asString _utName
 
-          invariants = schemas ^? ix schemaName.tMeta.mModel.ix "invariants"
-          invariant  = schemas ^? ix schemaName.tMeta.mModel.ix "invariant"
+    invariants <- case schemas ^? ix schemaName.tMeta.mModel._Just of
+      -- no model = no invariants
+      Nothing    -> pure []
+      Just model -> liftEither $ do
+        let model'        = expToMapping model
+            expInvariants = model' ^? _Just . ix "invariants"
+            expInvariant  = model' ^? _Just . ix "invariant"
+        exps <- collectExps "invariants" expInvariants expInvariant
+        runExpParserOver exps $
+          flip runReaderT (varIdArgs _utFields) . expToInvariant TBool
 
-          invInfo = runExpParserOver
-            "invariants" invariants invariant $
-            \meta -> runReaderT (expToInvariant TBool meta)
-              (varIdArgs _utFields)
+    pure $ Table tabName schema invariants
 
-      pure $ Table tabName schema <$> invInfo
-
-  let (failures, tables') = partitionEithers eitherTables
-  pure $ if null failures then Right tables' else Left (concat failures)
+-- | Parse a property definition like
+--
+-- * '(defproperty foo (> 1 0))'
+-- * '(defproperty foo (a:integer b:integer) (> a b))'
+parseDefprops :: Exp -> Either ParseFailure [(Text, DefinedProperty Exp)]
+parseDefprops (ELitList exps) = traverse parseDefprops' exps where
+  parseDefprops' exp@(EList' (EAtom' "defproperty" : rest)) = case rest of
+    [ EAtom' propname, EList' args, body ] -> do
+      args' <- parseBindings (curry Right) args & _Left %~ (exp,)
+      pure (propname, DefinedProperty args' body)
+    [ EAtom' propname,              body ] ->
+      pure (propname, DefinedProperty [] body)
+    _ -> Left (exp, "Invalid property definition")
+  parseDefprops' exp = Left (exp, "expected set of defproperty")
+parseDefprops exp = Left (exp, "expected set of defproperty")
 
 -- Get the set (HashMap) of refs to functions in this module.
 moduleTypecheckableRefs :: ModuleData -> HM.HashMap Text Ref
@@ -354,18 +383,24 @@ moduleTypecheckableRefs (_mod, modRefs) = flip HM.filter modRefs $ \case
   Ref (TConst {}) -> True
   _               -> False
 
+-- Get the set of properties defined in this module
+modulePropDefs
+  :: ModuleData -> Either ParseFailure (HM.HashMap Text (DefinedProperty Exp))
+modulePropDefs (Pact.Module{Pact._mMeta=Pact.Meta _ model}, _modRefs)
+  = case model of
+      Just model' -> HM.fromList <$> parseDefprops model'
+      Nothing     -> Right HM.empty
+
 moduleFunChecks
   :: [Table]
   -> HM.HashMap Text (Ref, Pact.FunType TC.UserType)
-  -> Either VerificationFailure
-       (HM.HashMap Text (Ref, Either [ParseFailure] [Located Check]))
-moduleFunChecks tables modTys = for modTys $
+  -> HM.HashMap Text (DefinedProperty Exp)
+  -> Except VerificationFailure
+       (HM.HashMap Text (Ref, Either ParseFailure [Located Check]))
+moduleFunChecks tables modTys propDefs = for modTys $
   \(ref@(Ref defn), Pact.FunType argTys resultTy) -> do
 
-  let properties = defn ^? tMeta . mModel . ix "properties"
-      property   = defn ^? tMeta . mModel . ix "property"
-
-      -- TODO: Ideally we wouldn't have any ad-hoc VID generation, but we're
+  let -- TODO: Ideally we wouldn't have any ad-hoc VID generation, but we're
       --       not there yet:
       vids = VarId <$> [1..]
 
@@ -402,19 +437,32 @@ moduleFunChecks tables modTys = for modTys $
                   (ColumnName (T.unpack argName),) <$> maybeTranslateType ty
           in (TableName (T.unpack _tableName), colMap)
 
-      checks :: Either [ParseFailure] [Located Check]
-      checks = runExpParserOver "properties" properties property
-        (expToCheck tableEnv vidStart nameEnv idEnv)
+  checks <- case defn ^? tMeta . mModel . _Just of
+    -- no model = no properties
+    Nothing    -> pure []
+    Just model -> withExcept ModuleParseFailure $ liftEither $ do
+      let model'        = expToMapping model
+          expProperties = model' ^? _Just . ix "properties"
+          expProperty   = model' ^? _Just . ix "property"
+      exps <- collectExps "properties" expProperties expProperty
+      runExpParserOver exps $
+        expToCheck tableEnv vidStart nameEnv idEnv propDefs
 
-  pure (ref, checks)
+  pure (ref, Right checks)
+
+-- | Given an exp like '(k v)', convert it to a singleton map
+expToMapping :: Exp -> Maybe (Map Text Exp)
+expToMapping (Pact.EList [Pact.EAtom k Nothing Nothing _, v] Nothing _)
+  = Just $ Map.singleton k v
+expToMapping _ = Nothing
 
 -- | For both properties and invariants you're allowed to use either the
 -- singular ("property") or plural ("properties") name. This helper just
 -- collects the properties / invariants in a list.
-collectExps :: String -> Maybe Exp -> Maybe Exp -> Either [(Exp, String)] [Exp]
+collectExps :: String -> Maybe Exp -> Maybe Exp -> Either ParseFailure [Exp]
 collectExps name multiExp singularExp = case multiExp of
   Just (Pact.ELitList exps') -> Right exps'
-  Just exp -> Left [(exp, name ++ " must be a list")]
+  Just exp -> Left (exp, name ++ " must be a list")
   Nothing -> case singularExp of
     Just exp -> Right [exp]
     Nothing  -> Right []
@@ -426,26 +474,12 @@ expToInfo exp = Info (Just (Code (tShow exp), exp ^. eParsed))
 -- or successes.
 runExpParserOver
   :: forall t.
-     String
-  -> Maybe Exp
-  -> Maybe Exp
+     [Exp]
   -> (Exp -> Either String t)
-  -> Either [ParseFailure] [Located t]
-runExpParserOver name multiExp singularExp parser =
-  let
-      exps = collectExps name multiExp singularExp
-      parsedList :: Either [ParseFailure] [Either [ParseFailure] (Located t)]
-      parsedList = exps <&&> \meta -> case parser meta of
-        Left err   -> Left [(meta, err)]
-        Right good -> Right (Located (expToInfo meta) good)
-
-      parsedList' :: [Either [ParseFailure] (Located t)]
-      parsedList' = join <$> sequence parsedList
-
-      failures :: [ParseFailure]
-      failures = join $ lefts parsedList'
-
-  in if null failures then Right (rights parsedList') else Left failures
+  -> Either ParseFailure [Located t]
+runExpParserOver exps parser = sequence $ exps <&> \meta -> case parser meta of
+  Left err   -> Left (meta, err)
+  Right good -> Right (Located (expToInfo meta) good)
 
 verifyFunctionProps :: [Table] -> Ref -> [Located Check] -> IO [CheckResult]
 verifyFunctionProps tables ref props = do
@@ -472,7 +506,7 @@ verifyFunctionInvariants tables ref = do
       if Set.null failures
       then verifyFunctionInvariants' _fInfo tables _fArgs _fBody
       else pure $ Left $ CheckFailure _fInfo (TypecheckFailure failures)
-    other -> pure $ Left $ CheckFailure (_tlInfo other) (NotAFunction (tShow ref))
+    _ -> pure $ Right $ TableMap Map.empty
 
 -- TODO: use from Control.Monad.Except when on mtl 2.2.2
 liftEither :: MonadError e m => Either e a -> m a
@@ -485,9 +519,30 @@ verifyModule
   -> ModuleData                         -- ^ the module we're verifying
   -> IO (Either VerificationFailure ModuleChecks)
 verifyModule modules moduleData = runExceptT $ do
-  tables <- withExceptT ModuleParseFailures $ moduleTables modules moduleData
+  tables <- withExceptT ModuleParseFailure $ moduleTables modules moduleData
 
-  let typecheckedRefs :: HM.HashMap Text Ref
+  let -- HM.unions is biased towards the start of the list. This module should
+      -- shadow the others. Note that load / shadow order of imported modules
+      -- is undefined and in particular not the same as their import order.
+      allModules = moduleData : HM.elems modules
+
+  allModulePropDefs <-
+    withExceptT ModuleParseFailure $ liftEither $
+      traverse modulePropDefs allModules
+
+  let -- how many times have these names been defined across all in-scope
+      -- modules
+      allModulePropNameDuplicates =
+          HM.keys
+        $ HM.filter (> (1 :: Int))
+        $ foldl (\acc k -> acc & at k %~ (Just . maybe 0 succ)) HM.empty
+        $ concatMap HM.keys
+        $ allModulePropDefs
+
+      propDefs :: HM.HashMap Text (DefinedProperty Exp)
+      propDefs = HM.unions allModulePropDefs
+
+      typecheckedRefs :: HM.HashMap Text Ref
       typecheckedRefs = moduleTypecheckableRefs moduleData
 
   -- For each ref, if it typechecks as a function (it'll be either a function
@@ -504,17 +559,17 @@ verifyModule modules moduleData = runExceptT $ do
     HM.empty
     typecheckedRefs
 
-  (funChecks :: HM.HashMap Text (Ref, Either [ParseFailure] [Located Check]))
-    <- liftEither $ moduleFunChecks tables funTypes
+  (funChecks :: HM.HashMap Text (Ref, Either ParseFailure [Located Check]))
+    <- hoist generalize $ moduleFunChecks tables funTypes propDefs
 
-  let funChecks' :: Either [ParseFailure] (HM.HashMap Text (Ref, [Located Check]))
+  let funChecks' :: Either ParseFailure (HM.HashMap Text (Ref, [Located Check]))
       funChecks' = traverse sequence funChecks
 
       verifyFunProps :: (Ref, [Located Check]) -> IO [CheckResult]
       verifyFunProps = uncurry (verifyFunctionProps tables)
 
   funChecks'' <- case funChecks' of
-    Left errs         -> throwError $ ModuleParseFailures errs
+    Left errs         -> throwError $ ModuleParseFailure errs
     Right funChecks'' -> pure funChecks''
 
   funChecks''' <- lift $ traverse verifyFunProps funChecks''
@@ -522,7 +577,9 @@ verifyModule modules moduleData = runExceptT $ do
     withExceptT ModuleCheckFailure $ ExceptT $
       verifyFunctionInvariants tables ref
 
-  pure $ ModuleChecks funChecks''' invariantChecks
+  let warnings = VerificationWarnings allModulePropNameDuplicates
+
+  pure $ ModuleChecks funChecks''' invariantChecks warnings
 
 -- | Verifies a one-off 'Check' for a function.
 verifyCheck
@@ -537,7 +594,7 @@ verifyCheck moduleData funName check = do
       moduleFun :: ModuleData -> Text -> Maybe Ref
       moduleFun (_mod, modRefs) name = name `HM.lookup` modRefs
 
-  tables <- withExceptT ModuleParseFailures $ moduleTables modules moduleData
+  tables <- withExceptT ModuleParseFailure $ moduleTables modules moduleData
   case moduleFun moduleData funName of
     Just funRef -> ExceptT $
       Right . head <$> verifyFunctionProps tables funRef [Located info check]
