+++ src/Pact/Repl/Lib.hs
@@ -359,16 +359,18 @@ verify i as = case as of
         modResult <- liftIO $ verifyModule modules md
         -- TODO: build describeModuleResult
         case modResult of
-          Left (ModuleParseFailures failures)  -> setop $ TcErrors $
-            fmap (Text.unpack . describeParseFailure) $ failures
-          Left (ModuleCheckFailure checkFailure) -> setop $ TcErrors $
-            pure $ Text.unpack $ describeCheckFailure checkFailure
-          Left (TypeTranslationFailure msg ty) -> setop $ TcErrors $ pure $
-            Text.unpack $ msg <> ": " <> tShow ty
-          Right (ModuleChecks propResults invariantResults) -> setop $ TcErrors $
+          Left (ModuleParseFailure failure)  -> setop $ TcErrors
+            [Text.unpack $ describeParseFailure failure]
+          Left (ModuleCheckFailure checkFailure) -> setop $ TcErrors
+            [Text.unpack $ describeCheckFailure checkFailure]
+          Left (TypeTranslationFailure msg ty) -> setop $ TcErrors
+            [Text.unpack $ msg <> ": " <> tShow ty]
+          Right (ModuleChecks propResults invariantResults warnings) -> setop $ TcErrors $
             let propResults'      = propResults      ^.. traverse . each
                 invariantResults' = invariantResults ^.. traverse . traverse . each
-            in Text.unpack . describeCheckResult <$> propResults' <> invariantResults'
+            in fmap Text.unpack $
+                 (describeCheckResult <$> propResults' <> invariantResults') <>
+                 [describeVerificationWarnings warnings]
 
         return (tStr "")
