+++ src/Pact/Types/Term.hs
@@ -79,7 +79,6 @@ import Control.DeepSeq
 import Data.Maybe
 import qualified Data.HashSet as HS
 import qualified Data.HashMap.Strict as HM
-import qualified Data.Map.Strict as M
 import Data.Int (Int64)
 import Data.Serialize (Serialize)
 
@@ -90,8 +89,8 @@ import Pact.Types.Exp
 
 
 data Meta = Meta
-  { _mDocs  :: !(Maybe Text)     -- ^ docs
-  , _mModel :: !(M.Map Text Exp) -- ^ model
+  { _mDocs  :: !(Maybe Text) -- ^ docs
+  , _mModel :: !(Maybe Exp)  -- ^ model
   } deriving (Eq, Show, Generic)
 instance ToJSON Meta where
   toJSON Meta {..} = object
@@ -225,7 +224,7 @@ instance ToJSON Module where
 -- address _mMeta not having a FromJSON, for now harmless
 instance FromJSON Module where
   parseJSON = withObject "Module" $ \o -> Module <$>
-    o .: "name" <*> o .: "keyset" <*> pure (Meta Nothing M.empty) {- o .:? "meta" -} <*>
+    o .: "name" <*> o .: "keyset" <*> pure (Meta Nothing Nothing) {- o .:? "meta" -} <*>
     o .: "code" <*> o .: "hash" <*> (HS.fromList <$> o .: "blessed")
 
 data ConstVal n =
